--- src/js/favorites/component/friends.module.css
@@ -1,3 +1,0 @@
-.friends{
-  margin-top: 0;
-}
