+++ src/js/menu/component/compatibility.jsx
@@ -1,11 +1,10 @@
 import React from "react";
 import css from "./compatibility.module.css";
-
 class Compatibility extends React.Component {
 
   render() {
     return (
-      <div>
+      <div className = {css.menu}>
         <button className = {css.vs_btn}>バトル相性診断</button>
       </div>
     );
