+++ src/js/menu/component/compatibility.module.css
@@ -13,3 +13,7 @@
     background-color: black;
     color: white;
 }
+.menu{
+    display: flex;
+    justify-content: center;
+}
