+++ src/js/menu/component/index.jsx
@@ -5,7 +5,7 @@ class Index extends React.Component {
 
   render() {
     return (
-      <div>
+      <div className = {css.menu}>
         <button className = {css.index_btn}>仲間ポケモン一覧</button>
       </div>
     );
