+++ src/js/menu/component/index.module.css
@@ -14,3 +14,8 @@
   background-color: black;
   color: white;
 }
+
+.menu{
+  display: flex;
+  justify-content: center;
+}
