+++ src/js/menu/component/records.module.css
@@ -14,3 +14,9 @@
   background-color: black;
   color: white;
 }
+
+.menu{
+  margin-top: 300px;
+  display: flex;
+  justify-content: center;
+}
