+++ src/js/records/component/record.module.css
@@ -1,0 +1,4 @@
+.record{
+  margin-top: 0 auto;
+  text-align: center;
+}
